Introduce io.vertx.core.net.Socket interface that abstracts the socket part of NetSocket which is a TCP socket
